pyproject.toml
[tool.poetry]
name = "arxiv-hunter"
version = "0.1.0"
description = ""
authors = ["Subrata Mondal <subratasubha2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6.1"
black = "^24.1.1"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pylint = "^3.0.3"
scikit-learn = "^1.4.0"
nltk = "^3.8.1"
pandas = "^2.2.0"
feedparser = "^6.0.11"
pymupdf = "^1.23.21"
python-dotenv = "^1.0.1"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
jinja2 = "^3.1.3"
requests = "^2.31.0"
sqlalchemy = "^2.0.27"
passlib = "^1.7.4"
psycopg2-binary = "^2.9.9"
sqlmodel = "^0.0.16"
asyncpg = "^0.29.0"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
python-jose = "^3.3.0"
pyjwt = "^2.8.0"
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
